+++ test/integration/production/test/index.test.js
@@ -95,7 +95,7 @@ describe('Production Usage', () => {
     expect(serverTrace.version).toBe(1)
     expect(
       serverTrace.files.some((file) =>
-        file.includes('next/dist/server/send-payload.js')
+        file.includes('next/dist/server/send-payload/index.js')
       )
     ).toBe(true)
     expect(
